level2

+++ crates/jam4/src/level/registry.rs
@@ -29,7 +29,7 @@ pub struct LevelInfo {
   pub starting_point: Vec2,
   pub next_level: Option<LevelId>,
   pub spawn_points: Vec<Vec2>,
-  pub boids_per_spawn_point: u8,
+  pub boids_per_spawn_point: i32,
   pub rescue_goal: Option<u16>,
   pub time_goal: Option<Duration>,
   pub wander: bool,
